feat: add Git plugin placeholder implementation
- Added plugin_git workspace member with GitConfig and GitOperations
- Implemented Plugin trait for GitPlugin with clone_repo, fetch, create_commit placeholders
- Prepares for WP-002 (Git integration) and Gix bindings
